requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
reportlab>=4.0.0
PyMuPDF>=1.23.0
cairosvg>=2.7.0
matplotlib>=3.7.0
numpy>=1.24.0
python-magic>=0.4.27
aiofiles>=23.0.0
Jinja2>=3.1.0